+++ Changelog.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Support `fossa test` and `fossa report` for monorepo projects ([#290](https://github.com/fossas/spectrometer/pull/290))
+- Maven `pom.xml`: Adds `${property}` substitution for `<groupId>` and `<artifactId>` fields in dependencies ([#282](https://github.com/fossas/spectrometer/pull/282))
 
 ## v2.10.3